% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_repo_database.R
\name{extract_repo_database}
\alias{extract_repo_database}
\title{Extract repo database}
\usage{
extract_repo_database(user, n_repos = 100)
}
\arguments{
\item{user}{GitHub User}

\item{n_repos}{Number of repos to fetch.}
}
\description{
Extract a database with the R packages hosted by an user on GH
}
